refactor: adicionando DTO na Classe Candidate
