+++ backend/src/main/groovy/repository/CandidateDAO.groovy
@@ -4,7 +4,7 @@ import model.Candidate
 import model.Skills
 import groovy.sql.GroovyRowResult
 import groovy.sql.Sql
-import groovy.transform.Canonical
+import repository.connection.DatabaseExecute
 import utils.OperationStatus
 
 import java.sql.SQLException
@@ -14,7 +14,7 @@ class CandidateDAO implements InterfaceDatabaseDML<Candidate> {
     private static final String SELECT_ALL_CANDIDATES = "SELECT * FROM candidatos AS c INNER JOIN usuarios AS u ON c.usuarios_id = u.id;";
     private static final String SELECT_CANDIDATE_SKILLS = "SELECT * FROM candidatos_competencias AS cc INNER JOIN competencias AS cs ON cc.competencias_id = cs.id WHERE cc.candidatos_id = ?;";
     private static final String SELECT_CANDIDATE_BY_ID = "SELECT * FROM candidatos AS c INNER JOIN usuarios AS u ON c.usuarios_id = u.id WHERE id = ?";
-    private static final String INSERT_CANDIDATE = "INSERT INTO candidatos (usuarios_id, nome, linkedin, data_nascimento) VALUES (?, ?, ?, ?)";
+    private static final String INSERT_CANDIDATE = "INSERT INTO candidatos (usuarios_id, nome, linkedin, data_nascimento, formacao) VALUES (?, ?, ?, ?, ?)";
     private static final String DELETE_CANDIDATE = "DELETE FROM candidatos WHERE usuarios_id = ?"
     private static final String UPDATE_CANDIDATE = "UPDATE candidatos SET nome = ?, sobrenome = ?, cpf = ?, descricao = ?, cidades_id = ?, cep = ?, formacao = ?, data_nascimento = ?, linkedin = ? WHERE usuarios_id = ?"
 
@@ -28,7 +28,7 @@ class CandidateDAO implements InterfaceDatabaseDML<Candidate> {
         List<Candidate> candidates = []
 
         sql.rows(SELECT_ALL_CANDIDATES).each { row ->
-            Candidate candidate = new Candidate(id: row.usuarios_id, name: row.nome, linkedin: row.linkedin)
+            Candidate candidate = new Candidate(id: row.usuarios_id, name: row.nome, cpf: row.cpf, description: row.descricao, city: row.cidades_id, academicEducation: row.formacao, linkedin: row.linkedin)
             candidates.add(candidate)
         }
         return candidates
@@ -46,14 +46,14 @@ class CandidateDAO implements InterfaceDatabaseDML<Candidate> {
     }
 
     Candidate findById(int id) throws SQLException {
-        GroovyRowResult result = sql.firstRow(SELECT_CANDIDATE_BY_ID, [id])
+        GroovyRowResult row = sql.firstRow(SELECT_CANDIDATE_BY_ID, [id])
 
-        Candidate candidate = new Candidate(id: result.id, name: result.nome, linkedin: result.linkedin)
+        Candidate candidate = new Candidate(id: row.usuarios_id, name: row.nome, cpf: row.cpf, description: row.descricao, city: row.cidades_id, academicEducation: row.formacao, linkedin: row.linkedin)
         return candidate
     }
 
     OperationStatus save(Candidate candidate) {
-        List<Object> arguments = [candidate.getId(), candidate.getName(), candidate.getLinkedin(), candidate.getDateOfBirth()]
+        List<Object> arguments = [candidate.getId(), candidate.getName(), candidate.getLinkedin(), candidate.getDateOfBirth(), candidate.getAcademicEducation()]
         return DatabaseExecute.executeTransaction(sql, INSERT_CANDIDATE, arguments)
     }
