+++ backend/src/main/groovy/repository/CompanyDAO.groovy
@@ -4,6 +4,7 @@ package repository
 import model.Company
 import groovy.sql.GroovyRowResult
 import groovy.sql.Sql
+import repository.connection.DatabaseExecute
 import utils.OperationStatus
 
 import java.sql.SQLException
