+++ backend/src/main/groovy/repository/JobsDAO.groovy
@@ -5,6 +5,7 @@ import model.Jobs
 import model.Skills
 import groovy.sql.GroovyRowResult
 import groovy.sql.Sql
+import repository.connection.DatabaseExecute
 import utils.OperationStatus
 
 import java.sql.SQLException
