+++ backend/src/main/groovy/repository/JobsSkillsDAO.groovy
@@ -4,9 +4,8 @@ package repository
 import model.Jobs
 import model.Skills
 import groovy.sql.Sql
+import repository.connection.DatabaseExecute
 import utils.OperationStatus
-
-import java.sql.SQLException
 
 class JobsSkillsDAO {
