+++ backend/src/main/groovy/repository/connection/DatabaseConfig.groovy
@@ -1,4 +1,4 @@
-package repository
+package repository.connection
 
 import groovy.sql.Sql
