+++ backend/src/main/groovy/repository/connection/DatabaseExecute.groovy
@@ -1,4 +1,4 @@
-package repository
+package repository.connection
 
 import groovy.sql.Sql
 import utils.OperationStatus
