+++ backend/src/main/groovy/repository/dto/CandidateDTO.groovy
@@ -1,0 +1,24 @@
+package repository.dto
+
+import model.Candidate
+import repository.CandidateDAO
+
+class CandidateDTO {
+
+    private final CandidateDAO candidateDAO
+
+    CandidateDTO(CandidateDAO candidateDAO) {
+        this.candidateDAO = candidateDAO
+    }
+
+    List<Candidate> findAll() {
+        List<Candidate> result = []
+
+        candidateDAO.findAll().forEach { candidate ->
+
+            result.add(new Candidate(id: candidate.getId(), academicEducation: candidate.getAcademicEducation()))
+        }
+
+        return result
+    }
+}
