+++ backend/src/test/groovy/model/CandidateTest.groovy
@@ -4,8 +4,7 @@ import groovy.sql.Sql
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import repository.CandidateDAO
-import repository.DatabaseConfig
-import repository.DatabaseSingleton
+import repository.connection.DatabaseSingleton
 import services.CandidateService
 
 import static org.junit.jupiter.api.Assertions.assertEquals
